restructure findr repo
