+++ test/extract-year-month-date.test.ts
@@ -4,7 +4,7 @@
  * This test suite validates the behavior of the extractYearMonthDate function when
  * extracting the year, month, and date from received message.
  */
-import { extractYearMonthDate } from '../src/adapters/extract-year-month-date';
+import { extractYearMonthDate } from '../src/utilities/extract-year-month-date';
 
 describe('extractYearMonthDate', () => {
   it('should extract year, month, and date from a timestamp', () => {
